Changed image-utils to use `writeFileSync` instead of piping to a stream.
Streams are async, and without returning the string,
a promise or using a callback, you cannot wait on
the stream response.

I'm using this in unit tests, and need to wait for
the image to be written before comparing it to
a snapshot image.

Should this need to be an async function, it can be
fixed differently.