Fixes #520. Warning: this is a substantial breaking change!
New matrices applied via the transform() functions of geom2, geom3, and
path2 should multiply with the new matrix on the left. That is, instead
of geom.transform = geom.transform * new_transform, it should be
geom.transform = new_transform * geom.transform.

Before this change, calls to transform() are effectively applied
Last-In-First-Out (stack order). This can lead to strange circumstances
like where a caller tries to move a shape long the X axis, but since
previous code already pushed a "rotate around the Z axis" matrix onto the
"stack", the call to translateX() actually translates along the Y axis,
which is clearly not the user's intent.

Fixed tests, and also changed them a bit to translate differently along
different axes. This makes the strangeness of the behavior more obvious.

This is a breaking change for any code that applies multiple transforms
to a single geometry (except for the special case where they commute).