fix(modeling): geom2 mirroring transform
Fixes #1301

When a mirroring transform is applied to a geometry, we should
reverse the point order so that orientation is preserved.

This was already done for geom3/poly3 geometries, but was broken for geom2.

This PR adds a check for negative 2D determinant to check if a transform
is mirroring in 2D.